requirements.txt
pillow==7.1.2
numpy==1.18.5
tqdm==4.46.1